add new commands in em

+++ src/batchling/experiment.py
@@ -13,8 +13,6 @@ from pydantic import (
     model_validator,
 )
 
-from batchling.db.crud import create_experiment, delete_experiment, update_experiment
-from batchling.db.session import get_db
 from batchling.request import (
     ProcessedRequest,
     RawRequest,
@@ -37,9 +35,8 @@ class Experiment(BaseModel, ABC):
         default="/v1/chat/completions",
         description="generation endpoint to use for the provider, e.g. /v1/chat/completions, /v1/embeddings..",
     )
-    api_key: str | None = Field(
-        default=None,
-        description="Optional, the API key to use for the provider if not using standard naming / env variables",
+    api_key: str = Field(
+        description="the API key to use for the provider if not using standard naming / env variables",
     )
     raw_requests: list[RawRequest] | None = Field(
         default_factory=list,
@@ -73,11 +70,11 @@ class Experiment(BaseModel, ABC):
             self.updated_at = now
         return self
 
-    @model_validator(mode="after")
-    def set_api_key(self):
-        if self.api_key is None:
-            self.api_key = get_default_api_key_from_provider(self.provider)
-        return self
+    @model_validator(mode="before")
+    def set_api_key(cls, values: dict) -> dict:
+        if values.get("api_key") is None:
+            values["api_key"] = get_default_api_key_from_provider(cls.provider)
+        return values
 
     @abstractmethod
     @cached_property
@@ -173,7 +170,6 @@ class Experiment(BaseModel, ABC):
         """Start the experiment:
         - create the processed file in the provider
         - create the batch in the provider
-        - update the database updated_at in the local db
 
         Returns:
             None
@@ -182,14 +178,6 @@ class Experiment(BaseModel, ABC):
             raise ValueError(f"Experiment in status {self.status} cannot be started")
         self.provider_file_id = self.create_provider_file()
         self.batch_id = self.create_provider_batch()
-        with get_db() as db:
-            update_experiment(
-                db=db,
-                id=self.id,
-                updated_at=datetime.now(),
-                batch_id=self.batch_id,
-                provider_file_id=self.provider_file_id,
-            )
 
     def cancel(self) -> None:
         """Cancel the experiment:
@@ -201,82 +189,44 @@ class Experiment(BaseModel, ABC):
         """
         self.raise_not_in_running_status()
         self.cancel_provider_batch()
-        with get_db() as db:
-            update_experiment(db=db, id=self.id, updated_at=datetime.now())
 
     def get_results(self):
-        os.makedirs(os.path.dirname(self.results_file_path), exist_ok=True)
         self.raise_not_in_completed_status()
+        os.makedirs(os.path.dirname(self.results_file_path), exist_ok=True)
         return self.get_provider_results()
-
-    def save(self):
-        if self.status != "created":
-            raise ValueError(f"Can only save an experiment in created status. Found: {self.status}")
-        with get_db() as db:
-            create_experiment(
-                db=db,
-                id=self.id,
-                model=self.model,
-                api_key=self.api_key,
-                name=self.name,
-                description=self.description,
-                provider=self.provider,
-                endpoint=self.endpoint,
-                raw_requests=self.raw_requests,
-                response_format=self.response_format,
-                processed_file_path=self.processed_file_path,
-                results_file_path=self.results_file_path,
-                created_at=self.created_at,
-                updated_at=self.updated_at,
-            )
-
-    def delete_local_experiment(self):
-        with get_db() as db:
-            delete_experiment(db=db, id=self.id)
-        if os.path.exists(self.processed_file_path):
-            os.remove(self.processed_file_path)
 
     def delete(self):
         """Delete:
-        - experiment from the database, if any
-        - local file, if any
         - provider file, if any
         - provider batch, if any
-        - provider output file, if any
-
         Returns:
             None
         """
-        self.delete_local_experiment()
+        if os.path.exists(self.processed_file_path):
+            os.remove(self.processed_file_path)
         if self.provider_file_id is not None:
             self.delete_provider_file()
         if self.batch_id is not None:
             self.delete_provider_batch()
 
     def update(self, **kwargs) -> "Experiment":
-        """Update the experiment by updating the database
+        """Update the experiment
 
         Returns:
             Experiment: The updated experiment
         """
-        if self.status != "created":
-            raise ValueError(
-                f"Can only update an experiment in created status. Found: {self.status}"
-            )
-        if "id" in kwargs:
-            raise ValueError(
-                "id cannot be updated, please delete the experiment and create a new one"
-            )
-        kwargs["updated_at"] = datetime.now()
-        exp_dict = self.model_dump()
-        exp_dict.update(kwargs)
+        updated_dict_experiment = self.model_dump()
+        updated_dict_experiment.update(kwargs)
         # validate model first to avoid updating the database with invalid data
-        exp_dict["raw_requests"] = [
-            RawRequest.model_validate(raw_request) for raw_request in exp_dict["raw_requests"]
+        updated_dict_experiment["raw_requests"] = [
+            RawRequest.model_validate(raw_request)
+            for raw_request in updated_dict_experiment["raw_requests"]
         ]
-        updated_experiment = self.__class__.model_validate(exp_dict)
-        with get_db() as db:
-            db_experiment = update_experiment(db=db, id=self.id, **kwargs)
-        if db_experiment is None:
-            raise ValueError(f"Experiment with id: {self.id} not found")
+        updated_experiment = self.__class__.model_validate(updated_dict_experiment)
+        if set(
+            ["raw_requests", "response_format", "endpoint", "model", "processed_file_path"]
+        ) & set(kwargs):
+            self.write_processed_batch_file()
+        if "processed_file_path" in kwargs and os.path.exists(self.processed_file_path):
+            os.remove(self.processed_file_path)
         return updated_experiment
